Use constants instead of package variables
We use namespace::clean to make sure the constants
are not available out of our package, except for
constants that were already being exposed.

Although namespace::clean raises the minimum version
of Perl for this distribution from 5.006 to 5.008,
PDL (on which we depend) already requires 5.010, so
this is not likely to cause any problems.