Tidy window tests
This patch re-writes the window tests quite fundamentally,
but the behaviour stays mostly the same. The biggest changes
are so that they are easier to read, and so that tests have
more informative names.

Additional changes make it so that tests that are impossible
are marked as skipped, and not simply skipped.