Drop dependency on namespace::clean
This also drops PDL::LiteF which means we now load only the parts of
PDL that we _actually_ use, and those can be immediately tracked
because PDL functions are used with their fully-qualified names.

This makes for code that is a little more cumbersome, but with the
benefit of being explicit and keeping PDL as our only dependency.